fix changes after review
