+++ QualityControl/test/setup/seeders/layout-show/json-file-mock.js
@@ -30,31 +30,3 @@ export const editedMockedLayout = {
   displayTimestamp: false,
   autoTabChange: 0,
 };
-
-export const defaultMockedLayout = {
-  name: 'a-test',
-  tabs: [
-    {
-      name: 'main',
-      objects: [
-        {
-          x: 0,
-          y: 0,
-          h: 1,
-          w: 1,
-          name: 'qc/test/object/1',
-          options: [],
-          ignoreDefaults: false,
-        },
-      ],
-      columns: 2,
-    },
-    {
-      name: 'a',
-      objects: [],
-      columns: 2,
-    },
-  ],
-  displayTimestamp: false,
-  autoTabChange: 0,
-};
